requirements.txt
boto3>=1.35.0
datasets>=2.19.0
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.32.3
tenacity>=8.2.3
tqdm>=4.66.4
